Fase 2 de desarrollo

+++ PrevAuto/src/domain/Alerta.java
@@ -5,12 +5,12 @@ import java.util.Date;
 public class Alerta {
 	
 	private String tipoDocumento;
-	private int numeroDocumento;
+	private Long numeroDocumento;
 	private String nombreVehiculo;
 	private String fechaVencimiento;
 	
 	
-	public Alerta(String tipoDocumento, int numeroDocumento, String nombreVehiculo, 
+	public Alerta(String tipoDocumento, Long numeroDocumento, String nombreVehiculo, 
 			String fechaVencimiento) {
 		this.tipoDocumento = tipoDocumento;
 		this.numeroDocumento = numeroDocumento;
